+++ src/renderer/components/composer/ComposerMessageInput.tsx
@@ -129,10 +129,10 @@ export default class ComposerMessageInput extends React.Component<
 
     // ENTER + SHIFT
     if (e.key === 'Enter' && e.shiftKey) {
-      return enterKeySends ? 'NEWLINE' : 'SEND'
+      return 'NEWLINE'
       // ENTER + CTRL
     } else if (e.key === 'Enter' && (e.ctrlKey || e.metaKey)) {
-      return enterKeySends ? 'NEWLINE' : 'SEND'
+      return 'SEND'
       // ENTER
     } else if (e.key === 'Enter' && !e.shiftKey) {
       return enterKeySends ? 'SEND' : 'NEWLINE'